Fix `total` variable/method overflowing 32-bits values

+++ src/main/java/ofc/bot/domain/entity/UserEconomy.java
@@ -62,7 +62,7 @@ public class UserEconomy extends OficinaRecord<UserEconomy> {
     }
 
     public long getTotal() {
-        return getWallet() + getBank();
+        return ((long) getWallet()) + ((long) getBank());
     }
 
     public long getLastDailyAt() {
